add brand&country

+++ AutoShow/AdminCrudForm.cs
@@ -48,6 +48,16 @@ namespace AutoShow
                                                                                            t.EngineLocation.EngineLocationName,
                                                                                            t.SeatsAmount}).ToList();
             }
+            else if(_option == Option.CarBrand)
+            {
+                HeaderLabel.Text = "Марки";
+                DataGridView.DataSource = _context.CarBrands.Select(c => new { c.CarBrandName }).ToList();
+            }
+            else if (_option == Option.Country)
+            {
+                HeaderLabel.Text = "Страны";
+                DataGridView.DataSource = _context.Countries.Select(c => new { c.CountryName }).ToList();
+            }
         }
 
         public void RefreshData()
@@ -75,6 +85,15 @@ namespace AutoShow
                     t.SeatsAmount
                 }).ToList();
             }
+            else if (_option == Option.CarBrand)
+            {
+                DataGridView.DataSource = _context.CarBrands.Select(c => new { c.CarBrandName }).ToList();
+            }
+            else if (_option == Option.Country)
+            {
+                HeaderLabel.Text = "Страны";
+                DataGridView.DataSource = _context.Countries.Select(c => new { c.CountryName }).ToList();
+            }
         }
 
         private void CloseLabel_Click(object sender, EventArgs e)
@@ -89,9 +108,10 @@ namespace AutoShow
             this.Close();
         }
 
-        private void AddButton_Click(object sender, EventArgs e)
+        private void CreateButton_Click(object sender, EventArgs e)
         {
-            if (_option == Option.BodyType || _option == Option.EngineType || _option == Option.EngineLocation)
+            if (_option == Option.BodyType || _option == Option.EngineType || _option == Option.EngineLocation || _option == Option.CarBrand ||
+                _option == Option.Country)
             {
                 var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option);
                 createUpdateAdminForm.Show();
@@ -99,7 +119,7 @@ namespace AutoShow
             else if(_option == Option.TechnicalInformation)
             {
                 var createUpdateAdminForm = new CU_TechnicalInformationForm(_context, this);
-                createUpdateAdminForm.Show(); 
+                createUpdateAdminForm.Show();
             }
             this.Hide();
         }
@@ -151,6 +171,20 @@ namespace AutoShow
                 var technicalInforamtion = _context.TechnicalInformations.FirstOrDefault(t => t.BodyTypeId == bodyTypeId && t.EngineTypeId == engineTypeId &&
                 t.EngineDisplacement == engineDisplacement && t.DoorsAmount == doorsAmount && t.EngineLocationId == engineLocationId && t.SeatsAmount == seatsAmount);
                 var createUpdateAdminForm = new CU_TechnicalInformationForm(_context, this, technicalInforamtion);
+                createUpdateAdminForm.Show();
+            }
+            else if(_option == Option.CarBrand)
+            {
+                string carBrandName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                var carBrand = _context.CarBrands.FirstOrDefault(c => c.CarBrandName == carBrandName);
+                var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option, carBrand);
+                createUpdateAdminForm.Show();
+            }
+            else if (_option == Option.Country)
+            {
+                string countryName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                var country = _context.Countries.FirstOrDefault(c => c.CountryName == countryName);
+                var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option, country);
                 createUpdateAdminForm.Show();
             }
             this.Hide();
@@ -208,12 +242,46 @@ namespace AutoShow
                 string engineLocationName = DataGridView[4, DataGridView.SelectedRows[0].Index].Value.ToString();
                 int bodyTypeId = _context.BodyTypes.FirstOrDefault(b => b.BodyTypeName == bodyTypeName).BodyTypeId;
                 int engineTypeId = _context.EngineTypes.FirstOrDefault(en => en.EngineTypeName == engineTypeName).EngineTypeId;
-                int EngineDisplacement = int.Parse(DataGridView[2, DataGridView.SelectedRows[0].Index].Value.ToString());
+                int engineDisplacement = int.Parse(DataGridView[2, DataGridView.SelectedRows[0].Index].Value.ToString());
                 int doorsAmount = int.Parse(DataGridView[3, DataGridView.SelectedRows[0].Index].Value.ToString());
                 int engineLocationId = _context.EngineLocations.FirstOrDefault(en => en.EngineLocationName == engineLocationName).EngineLocationId;
                 int seatsAmount = int.Parse(DataGridView[5, DataGridView.SelectedRows[0].Index].Value.ToString());
+                //протестить эту часть
+                var carModel = _context.CarModels.Where(c => c.TechnicalInformation.BodyTypeId == bodyTypeId && 
+                c.TechnicalInformation.EngineTypeId == engineTypeId && c.TechnicalInformation.EngineDisplacement == engineDisplacement &&
+                c.TechnicalInformation.DoorsAmount == doorsAmount && c.TechnicalInformation.EngineLocationId == engineLocationId &&
+                c.TechnicalInformation.SeatsAmount == seatsAmount).FirstOrDefault();
+                if (carModel != null)
+                {
+                    MessageBox.Show("Данная техническая информация присутствует в моделях,поэтому для начала удалите все связанные данные");
+                    return;
+                }
                 _context.TechnicalInformations.Remove(_context.TechnicalInformations.FirstOrDefault(t => t.BodyTypeId == bodyTypeId && t.EngineTypeId == engineTypeId &&
-                t.EngineDisplacement == EngineDisplacement && t.DoorsAmount == doorsAmount && t.EngineLocationId == engineLocationId && t.SeatsAmount == seatsAmount));
+                t.EngineDisplacement == engineDisplacement && t.DoorsAmount == doorsAmount && t.EngineLocationId == engineLocationId && t.SeatsAmount == seatsAmount));
+            }
+            else if(_option == Option.CarBrand)
+            {
+                string carBrandName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                //и это протестить
+                var carModel = _context.CarModels.Where(c => c.CarBrand.CarBrandName == carBrandName).FirstOrDefault();
+                if (carModel != null)
+                {
+                    MessageBox.Show("Данная марка присутсвует в моделях,поэтому для начала удалите все связанные данные");
+                    return;
+                }
+                _context.CarBrands.Remove(_context.CarBrands.FirstOrDefault(c => c.CarBrandName == carBrandName));
+            }
+            else if (_option == Option.Country)
+            {
+                string countryName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                //и это протестить
+                var carModel = _context.CarModels.Where(c => c.Country.CountryName == countryName).FirstOrDefault();
+                if (carModel != null)
+                {
+                    MessageBox.Show("Данная страна присутсвует в моделях,поэтому для начала удалите все связанные данные");
+                    return;
+                }
+                _context.Countries.Remove(_context.Countries.FirstOrDefault(c => c.CountryName == countryName));
             }
             if (_context.SaveChanges() > 0)
             {
